# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TripFit API",
    "version": "v0.0.1"
  },
  "servers": [
    {
      "url": "https://api.tripfit.online",
      "description": "Production API Server"
    }
  ],
  "security": [
    {
      "bearer-jwt": []
    }
  ],
...
  "paths": {
...
    "/api/v1/trips/{tripId}": {
      "get": {
        "tags": [
          "Trip"
        ],
        "summary": "여행방 상세",
        "description": "목적: 여행방 상세 정보를 조회한다.\n\n호출 시점: 방 홈·설정 화면 진입.\n\n전제: 멤버이며 ACTIVE(일정 confirm/join 완료)이고 방 입장 조건(일정≥1 또는 전부 free)을 충족.\n\n결과: TripDetailResponse(inviteCode 포함 — 방장 초대 공유용).\n\n주의: SCHEDULE_PENDING(방장 confirm 전)면 SCHEDULE_CONFIRM_REQUIRED. create 응답에는 inviteCode가 없다.\n\n주요 에러: TRIP_ACCESS_DENIED · SCHEDULE_CONFIRM_REQUIRED · SCHEDULE_ENTRY_REQUIRED\n",
        "operationId": "getTrip",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseTripDetailResponse"
                },
                "example": {
                  "data": {
                    "tripId": "550e8400-e29b-41d4-a716-446655440000",
                    "name": "제주도 여행",
                    "destination": "제주도",
                    "startRange": "2026-08-01",
                    "endRange": "2026-08-31",
                    "durationDays": 4,
                    "durationNights": 3,
                    "memberCount": 6,
                    "status": "ONGOING",
                    "inviteCode": "AB12CD",
                    "confirmedStartDate": null,
                    "confirmedEndDate": null,
                    "lastRecommendationMode": null,
                    "lastActivityAt": "2026-07-20T10:00:00",
                    "pinned": false,
                    "myRole": "OWNER",
                    "myMemberStatus": "ACTIVE",
                    "activeMemberCount": 3,
                    "memberFillRate": 0.5,
                    "membersPreview": [
                      {
                        "userId": "550e8400-e29b-41d4-a716-446655440000",
                        "profileImageUrl": "https://lh3.googleusercontent.com/a/example",
                        "role": "OWNER"
                      }
                    ],
                    "membersPreviewOverflow": 0
                  }
                }
              }
            }
          },
          "401": {
            "description": "액세스 토큰 없음·무효(AUTH_INVALID_TOKEN)·만료(AUTH_EXPIRED)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_EXPIRED",
                  "message": "액세스 토큰이 만료되었습니다."
                }
              }
            }
          },
          "403": {
            "description": "TRIP_ACCESS_DENIED — 비참여자 · SCHEDULE_CONFIRM_REQUIRED — 이 방 일정 확인 미완료 · SCHEDULE_ENTRY_REQUIRED — 입장 조건 미충족",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "TRIP_ACCESS_DENIED",
                  "message": "여행방 참여 권한이 없습니다."
                }
              }
            }
          },
          "404": {
            "description": "TRIP_NOT_FOUND — 여행방 없음·soft deleted",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "TRIP_NOT_FOUND",
                  "message": "여행방을 찾을 수 없습니다."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Trip"
        ],
        "summary": "여행방 삭제",
        "description": "목적: 여행방을 삭제(soft)한다.\n\n호출 시점: 방장이 방 삭제 확인.\n\n전제: 방장. SCHEDULE_PENDING(confirm 전)여도 삭제 가능 — 단, 삭제는 메타 권한이지 방 입장·공유가 아님.\n\n결과: 204 No Content. 멤버 row도 연쇄 soft delete.\n\n주요 에러: TRIP_FORBIDDEN — 방장 아님\n",
        "operationId": "deleteTrip",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "삭제 성공(No Content)"
          },
          "401": {
            "description": "액세스 토큰 없음·무효(AUTH_INVALID_TOKEN)·만료(AUTH_EXPIRED)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_EXPIRED",
                  "message": "액세스 토큰이 만료되었습니다."
                }
              }
            }
          },
          "403": {
            "description": "TRIP_FORBIDDEN — 방장 아님",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "TRIP_FORBIDDEN",
                  "message": "여행방 방장만 수행할 수 있습니다."
                }
              }
            }
          },
          "404": {
            "description": "TRIP_NOT_FOUND — 여행방 없음·soft deleted",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "TRIP_NOT_FOUND",
                  "message": "여행방을 찾을 수 없습니다."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Trip"
        ],
        "summary": "여행방 메타 수정",
        "description": "목적: 방 이름·인원·여행지 등 메타를 수정한다.\n\n호출 시점: 방 설정에서 저장.\n\n전제: 방장. 여행방이 ONGOING(조율 중). 희망 기간(startRange~endRange)은 수정 불가.\n\n결과: 갱신된 TripDetailResponse.\n\n주요 에러: TRIP_FORBIDDEN — 방장 아님 · TRIP_NOT_ONGOING — 조율 중이 아님\n",
        "operationId": "patchTrip",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchTripRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "수정 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseTripDetailResponse"
                },
                "example": {
                  "data": {
                    "tripId": "550e8400-e29b-41d4-a716-446655440000",
                    "name": "제주도 여행",
                    "destination": "제주도",
                    "startRange": "2026-08-01",
                    "endRange": "2026-08-31",
                    "durationDays": 4,
                    "durationNights": 3,
                    "memberCount": 6,
                    "status": "ONGOING",
                    "inviteCode": "AB12CD",
                    "confirmedStartDate": null,
                    "confirmedEndDate": null,
                    "lastRecommendationMode": null,
                    "lastActivityAt": "2026-07-20T10:00:00",
                    "pinned": false,
                    "myRole": "OWNER",
                    "myMemberStatus": "ACTIVE",
                    "activeMemberCount": 3,
                    "memberFillRate": 0.5,
                    "membersPreview": [
                      {
                        "userId": "550e8400-e29b-41d4-a716-446655440000",
                        "profileImageUrl": "https://lh3.googleusercontent.com/a/example",
                        "role": "OWNER"
                      }
                    ],
                    "membersPreviewOverflow": 0
                  }
                }
              }
            }
          },
          "400": {
            "description": "요청 값 검증 실패 (INVALID_INPUT)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "INVALID_INPUT",
                  "message": "입력값이 올바르지 않습니다.",
                  "errors": [
                    {
                      "field": "name",
                      "message": "이름은 최대 15자입니다."
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "액세스 토큰 없음·무효(AUTH_INVALID_TOKEN)·만료(AUTH_EXPIRED)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_EXPIRED",
                  "message": "액세스 토큰이 만료되었습니다."
                }
              }
            }
          },
          "403": {
            "description": "TRIP_FORBIDDEN — 방장 아님",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "TRIP_FORBIDDEN",
                  "message": "여행방 방장만 수행할 수 있습니다."
                }
              }
            }
          },
          "404": {
            "description": "TRIP_NOT_FOUND — 여행방 없음·soft deleted",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "TRIP_NOT_FOUND",
                  "message": "여행방을 찾을 수 없습니다."
                }
              }
            }
          },
          "409": {
            "description": "TRIP_NOT_ONGOING — 조율 중이 아닌 여행방",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "TRIP_NOT_ONGOING",
                  "message": "조율 중인 여행방만 수정·내보내기·일정 확인할 수 있습니다."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ]
      }
    },
...
    "/api/v1/trips/{tripId}/members": {
      "get": {
        "tags": [
          "Trip Members"
        ],
        "summary": "참여자 목록",
        "description": "목적: 여행방 참여자 목록을 조회한다.\n\n호출 시점: 방 멤버 화면.\n\n전제: 멤버이며 방 입장 가능(ACTIVE + 입장 조건). SCHEDULE_PENDING면 403.\n\n결과: 상태·역할·Pin·모집률 등. 동명이인은 표시명에 `(2)`처럼 번호가 붙는다.\n",
        "operationId": "listMembers",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseTripMembersResponse"
                },
                "example": {
                  "data": {
                    "memberCount": 6,
                    "activeMemberCount": 3,
                    "memberFillRate": 0.5,
                    "members": [
                      {
                        "userId": "550e8400-e29b-41d4-a716-446655440000",
                        "displayName": "홍길동",
                        "role": "OWNER",
                        "memberStatus": "ACTIVE",
                        "pinned": true
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "액세스 토큰 없음·무효(AUTH_INVALID_TOKEN)·만료(AUTH_EXPIRED)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_EXPIRED",
                  "message": "액세스 토큰이 만료되었습니다."
                }
              }
            }
          },
          "403": {
            "description": "TRIP_ACCESS_DENIED — 비참여자 · SCHEDULE_CONFIRM_REQUIRED — 이 방 일정 확인 미완료(SCHEDULE_PENDING) · SCHEDULE_ENTRY_REQUIRED — 입장 조건 미충족",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "SCHEDULE_CONFIRM_REQUIRED",
                  "message": "이 여행방 일정 확인을 완료해야 입장할 수 있습니다."
                }
              }
            }
          },
          "404": {
            "description": "TRIP_NOT_FOUND — 여행방 없음·soft deleted",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "TRIP_NOT_FOUND",
                  "message": "여행방을 찾을 수 없습니다."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ]
      }
    },
    "/api/v1/trips/{tripId}/members/schedule-calendar": {
      "get": {
        "tags": [
          "Trip Members"
        ],
        "summary": "멤버 effective 일정 달력",
        "description": "목적: 희망 기간(startRange~endRange) 동안 멤버 전원의 effective 일정을 조회한다.\n\n호출 시점: 방 안 일정 조율·추천 전 달력.\n\n전제: 멤버이며 방 입장 가능.\n\n결과: 멤버별 날짜 슬롯. 조율 중(ONGOING)은 실시간 일정, 확정·종료 방은 당시 스냅샷(읽기 전용).\n\n주요 에러: TRIP_ACCESS_DENIED / SCHEDULE_CONFIRM_REQUIRED\n",
        "operationId": "getScheduleCalendar",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseMemberScheduleCalendarResponse"
                },
                "example": {
                  "data": {
                    "startDate": "2026-08-01",
                    "endDate": "2026-08-31",
                    "readOnly": false,
                    "members": [
                      {
                        "userId": "550e8400-e29b-41d4-a716-446655440000",
                        "displayName": "홍길동",
                        "role": "OWNER",
                        "memberStatus": "ACTIVE",
                        "days": [
                          {
                            "date": "2026-08-03",
                            "morningStatus": "IMPOSSIBLE",
                            "afternoonStatus": "IMPOSSIBLE",
                            "eveningStatus": "POSSIBLE",
                            "uncertain": false
                          }
                        ]
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "액세스 토큰 없음·무효(AUTH_INVALID_TOKEN)·만료(AUTH_EXPIRED)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_EXPIRED",
                  "message": "액세스 토큰이 만료되었습니다."
                }
              }
            }
          },
          "403": {
            "description": "TRIP_ACCESS_DENIED — 비참여자 · SCHEDULE_CONFIRM_REQUIRED — 이 방 일정 확인 미완료",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "TRIP_ACCESS_DENIED",
                  "message": "여행방 참여 권한이 없습니다."
                }
              }
            }
          },
          "404": {
            "description": "TRIP_NOT_FOUND — 여행방 없음·soft deleted",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "TRIP_NOT_FOUND",
                  "message": "여행방을 찾을 수 없습니다."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ]
      }
    },
...
    "/api/v1/trips/{tripId}/members/me": {
      "delete": {
        "tags": [
          "Trip Members"
        ],
        "summary": "여행방 나가기",
        "description": "목적: 참여자(MEMBER)가 스스로 여행방에서 나간다.\n\n호출 시점: 마이페이지·방 안 메뉴에서 나가기 확인.\n\n전제: 호출자가 이 방의 활성 멤버(MEMBER)다. 방장은 사용할 수 없다.\n\n결과: 본인 참여 기록이 soft delete되고 204를 반환한다. 방 상태(ONGOING/CONFIRMED/EXPIRED)와 무관하게 항상 허용된다.\n\n주의: 같은 초대 코드로 다시 참여할 수 있다. 방장은 이 API 대신 여행방 삭제를 사용해야 한다.\n\n주요 에러: TRIP_ACCESS_DENIED — 비참여자 또는 이미 나감 · TRIP_OWNER_CANNOT_LEAVE — 방장\n",
        "operationId": "leaveTrip",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "나가기 성공(No Content)"
          },
          "400": {
            "description": "TRIP_OWNER_CANNOT_LEAVE — 방장은 나갈 수 없음(방 삭제 사용)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "TRIP_OWNER_CANNOT_LEAVE",
                  "message": "방장은 여행방을 나갈 수 없습니다. 여행방 삭제를 이용해주세요."
                }
              }
            }
          },
          "401": {
            "description": "액세스 토큰 없음·무효(AUTH_INVALID_TOKEN)·만료(AUTH_EXPIRED)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_EXPIRED",
                  "message": "액세스 토큰이 만료되었습니다."
                }
              }
            }
          },
          "403": {
            "description": "TRIP_ACCESS_DENIED — 비참여자 또는 이미 나감",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "TRIP_ACCESS_DENIED",
                  "message": "여행방 참여 권한이 없습니다."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "API 에러 응답 envelope",
        "properties": {
          "code": {
            "type": "string",
            "description": "에러 코드",
            "example": "AUTH_INVALID_TOKEN"
          },
          "message": {
            "type": "string",
            "description": "사용자용 에러 메시지",
            "example": "유효하지 않은 토큰입니다."
          },
          "errors": {
            "type": [
              "array",
              "null"
            ],
            "description": "필드 검증 오류 목록 (validation 실패 시)",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
...
      "SuccessResponseTripDetailResponse": {
        "type": "object",
        "description": "API 성공 응답 envelope",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TripDetailResponse",
            "description": "응답 본문"
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "에러 메시지 (성공 시 null)"
          },
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "에러 코드 (성공 시 null)"
          }
        }
      },
...
      "PatchTripRequest": {
        "type": "object",
        "description": "여행방 메타 수정 요청. PATCH /trips/{tripId} (방장·ONGOING만). 희망 기간은 수정 불가",
        "properties": {
          "name": {
            "type": "string",
            "description": "여행방 이름 (최대 15자)",
            "example": "제주 3박4일",
            "maxLength": 15,
            "minLength": 1
          },
          "durationNights": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "희망 여행 박수 (n박). durationDays와 둘 다 null=미정, 둘 다 값이면 nights+1 ≤ days ≤ nights+2. 0박(당일치기)도 동일 규칙 적용(days 1~2)",
            "example": 3
          },
          "durationDays": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "희망 여행 일수 (m일). null=미정. durationNights+1 ~ durationNights+2 범위만 허용",
            "example": 4
          },
          "memberCount": {
            "type": "integer",
            "format": "int32",
            "description": "모집 정원 (1~10)",
            "example": 6,
            "maximum": 10,
            "minimum": 1
          },
          "destination": {
            "type": [
              "string",
              "null"
            ],
            "description": "여행지. null=미정",
            "example": "제주"
          }
        },
        "required": [
          "memberCount",
          "name"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer-jwt": {
        "type": "http",
        "description": "Authorization: Bearer {accessToken}",
        "name": "bearer-jwt",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}